feat: refactor subscription repository structure and implement ISubscriptionRepository methods
