--- ZucchiniMVC/Infrastructure/Repositories/ISubscriptionRepository.cs
@@ -1,9 +1,0 @@
-using ZucchiniCore.Entities;
-namespace Infrastrcture.Repositories;
-
-public interface ISubscriptionRepository
-{
-    Task AddSubscriptionAsync(Subscription subscription);
-    Task<SubscriptionType?> FindSubscriptionTypeByIdAsync(int id);
-    Task<List<SubscriptionType>> GetAllSubscriptionTypesAsync();
-}
