+++ ZucchiniMVC/Infrastructure/Repositories/SubscriptionRepo/ISubscriptionRepository.cs
@@ -1,0 +1,12 @@
+using ZucchiniCore.Entities;
+namespace Infrastrcture.Repositories.SubscriptionRepo
+{
+    public interface ISubscriptionRepository
+    {
+        Task AddSubscriptionAsync(Subscription subscription);
+        Task<SubscriptionType?> FindSubscriptionTypeByIdAsync(int id);
+        Task<List<SubscriptionType>> GetAllSubscriptionTypesAsync();
+    }
+
+}
+
